requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
config==0.4.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.2
cryptography==2.5
decorator==4.3.2
dj-database-url==0.5.0
Django==2.1.7
django-bootstrap3==11.0.0
django-extensions==2.1.5
django-filter==2.1.0
django-heroku==0.3.1
django-multiselectfield==0.1.8
django-oauth-plus==2.2.9
django-oauth-toolkit==1.2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
djangorestframework-stubs==0.2.0
Faker==0.9.1
gunicorn==19.6.0
httplib2==0.12.1
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
mixer==6.1.3
more-itertools==6.0.0
oauth2==1.9.0.post1
oauthlib==3.0.1
paramiko==2.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.9.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
PyNaCl==1.3.0
pytest==4.3.0
pytest-cov==2.6.1
pytest-django==3.4.8
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.0.0
requests==2.21.0
six==1.12.0
soupsieve==1.8
storage==0.0.4.3
text-unidecode==1.2
tornado==5.1.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
whitenoise==4.1.2
